changes for complete mock catalog processing

+++ py/LSS/common_tools.py
@@ -561,7 +561,7 @@ def addnbar(fb,nran=18,bs=0.01,zmin=0.01,zmax=1.6,P0=10000,add_data=True,ran_sw=
     mean_comp = len(fd)/np.sum(fd['WEIGHT_COMP'])
     print('mean completeness '+str(mean_comp))
     if 'NTILE' not in list(fd.dtype.names):
-        fd['NTILE'] = np.ones(len(fd))
+        fd['NTILE'] = np.ones(len(fd),dtype=int)
         print('added NTILE = 1 column because column did not exist')
     ntl = np.unique(fd['NTILE'])
     comp_ntl = np.ones(len(ntl))
@@ -628,7 +628,7 @@ def addnbar(fb,nran=18,bs=0.01,zmin=0.01,zmax=1.6,P0=10000,add_data=True,ran_sw=
         #fd['NZ'] = nl
         print('mean completeness '+str(mean_comp))
         if 'NTILE' not in list(fd.dtype.names):
-            fd['NTILE'] = np.ones(len(fd))
+            fd['NTILE'] = np.ones(len(fd),dtype=int)
             print('added NTILE = 1 column because column did not exist')
 
         fd['NX'] = nl*comp_ntl[fd['NTILE']-1]
